Move instance! macro from ipasir-sys to ipasir
`ipasir-sys` crate should not have a responsibility to support external libraries that implement IPASIR. Actually, `minisat` and `glucose` libraries (libminisat.so etc) anyway do not have ipasir-related code inside, so it was an error to load `minisat` as IPASIR-compliant ffi instance — when it was not!
Now, the USER is responsible of providing a library which compiles/links with IPASIR. For example, `cadical` library (libcadical.so) includes ccadical.o, which is IPASIR-compliant, so we provide an instance (via ipasir_instance! macro) of it in `ipasir` crate — note that we are doing this "one step away (above)" from the `ipasir-sys` crate.