Suppress clippy warning about Expr::not
Clippy suggests (`clippy::should_implement_trait`) implementing ops::Not or choosing less ambiguous name. However, Expr already implements ops::Not. Moreover, Expr::not is an associated function, not a &self method (actually, I think this is the reason why clippy does not "see" the existing ops::Not impl), so it is not possible to choose a different name for this constructor.
Overall, this is completely safe to ignore this clippy warning.